requirements.txt
dill==0.3.7
faiss-cpu==1.7.4
Flask==2.1.3
Flask-RESTful==0.3.10
numpy==1.23.1
opencv-python==4.6.0.66
ultralytics==8.0.203
Werkzeug==2.1.2
